pyproject.toml
[project]
name = "amoved"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "audio-recorder-streamlit>=0.0.10",
    "boto3>=1.38.44",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.0",
]
